Features: 0.5.0 - Add caching of -Path packages and add support for SemVer2 packages

+++ Import-Package/Import-Package.psm1
@@ -6,6 +6,9 @@ $loaded = @{
 }
 $mutexes = @{}
 
+New-Item (Join-Path $PSScriptRoot "Packages") -Force -ItemType Directory
+New-Item (Join-Path $PSScriptRoot "Temp") -Force -ItemType Directory
+
 & {
     # Clear the Cache
     Write-Verbose "[Import-Package:Init] Clearing Temp Files..."
@@ -21,6 +24,7 @@ $mutexes = @{}
 }
 
 
+. "$PSScriptRoot\src\ConvertTo-SemVerObject.ps1"
 . "$PSScriptRoot\src\Resolve-DependencyVersions.ps1"
 . "$PSScriptRoot\src\Resolve-CachedPackage.ps1"
 . "$PSScriptRoot\src\Build-PackageData.ps1"
@@ -174,44 +178,52 @@ function Import-Package {
         [string] $Path,
         
         [switch] $Offline,
-        # [string] $CachePath = "$PSScriptRoot\Packages"
-        [string] $TempPath = (& {
-            $parent = & {
-                [System.IO.Path]::GetTempPath()
-                # Join-Path ($CachePath | Split-Path -Parent) "Temp"
-            }
-            [string] $uuid = [System.Guid]::NewGuid()
-
-            # Cut dirname in half by compressing the UUID from base16 (hexadecimal) to base36 (alphanumeric)
-            $id = & {
-                $bigint = [uint128]::Parse( $uuid.ToString().Replace("-",""), 'AllowHexSpecifier')
-                $compressed = ""
-                
-                # Make hex-string more compressed by encoding it in base36 (alphanumeric)
-                $chars = "0123456789abcdefghijklmnopqrstuvwxyz"
-                While( $bigint -gt 0 ){
-                    $remainder = $bigint % 36
-                    $compressed = $chars[$remainder] + $compressed
-                    $bigint = $bigint/36
-                }
-                Write-Verbose "[Import-Package:Directories] UUID $uuid (base16) converted to $compressed (base$( $chars.Length ))"
-
-                $compressed
-            }
-
-            $mutexes."$id" = New-Object System.Threading.Mutex($true, "Global\ImportPackage-$id") # Lock the directory from automatic removal
-
-            New-Item -ItemType Directory -Path (Join-Path $parent $id) -Force
-            # Resolve-Path "."
-        })
+        [string] $CachePath = "$PSScriptRoot\Packages",
+        [string] $TempPath
     )
 
     Process {
+
+        $temp_path_generated = If( [string]::IsNullOrWhiteSpace( $TempPath ) ){
+            $TempPath = & {
+                $parent = & {
+                    Join-Path ($CachePath | Split-Path -Parent) "Temp"
+                }
+                [string] $uuid = [System.Guid]::NewGuid()
+    
+                # Cut dirname in half by compressing the UUID from base16 (hexadecimal) to base36 (alphanumeric)
+                $id = & {
+                    $bigint = [uint128]::Parse( $uuid.ToString().Replace("-",""), 'AllowHexSpecifier')
+                    $compressed = ""
+                    
+                    # Make hex-string more compressed by encoding it in base36 (alphanumeric)
+                    $chars = "0123456789abcdefghijklmnopqrstuvwxyz"
+                    While( $bigint -gt 0 ){
+                        $remainder = $bigint % 36
+                        $compressed = $chars[$remainder] + $compressed
+                        $bigint = $bigint/36
+                    }
+                    Write-Verbose "[Import-Package:Preparation] UUID $uuid (base16) converted to $compressed (base$( $chars.Length ))"
+    
+                    $compressed
+                }
+    
+                $mutexes."$id" = New-Object System.Threading.Mutex($true, "Global\ImportPackage-$id") # Lock the directory from automatic removal
+    
+                New-Item -ItemType Directory -Path (Join-Path $parent $id) -Force
+                # Resolve-Path "."
+            }
+            $true
+        } Else {
+            $false
+        }
+
         $PackageData = Switch( $PSCmdlet.ParameterSetName ){
             "Managed-Object" {
                 Write-Verbose "[Import-Package:ParameterSet] Managed Object"
 
                 Build-PackageData -From "Object" -Options @( $Package, @{
+                    "CachePath" = $CachePath
                     "TempPath" = $TempPath
                 }) -Bootstrapper $bootstrapper
             }
@@ -219,6 +231,7 @@ function Import-Package {
                 Write-Verbose "[Import-Package:ParameterSet] Managed"
 
                 Build-PackageData -From "Install" -Options @{
+                    "CachePath" = $CachePath
                     "TempPath" = $TempPath
 
                     "Offline" = $Offline # If true, do not install
@@ -231,6 +244,7 @@ function Import-Package {
                 Write-Verbose "[Import-Package:ParameterSet] Unmanaged"
 
                 Build-PackageData -From "File" -Options @{
+                    "CachePath" = $CachePath
                     "TempPath" = $TempPath
 
                     "Source" = $Path
@@ -333,9 +347,9 @@ function Import-Package {
                     } Else {
                         Write-Verbose "[Import-Package:Dependency-Handling] ($($PackageData.Name) Dependency) Loading $($_.Name) - $($_.Version) (Framework $( $package_framework.GetShortFolderName() ))"
                         If( $PackageData.Offline ){
-                            Import-Package $_.Name -Version $_.Version -TargetFramework $package_framework -Offline
+                            Import-Package $_.Name -Version $_.Version -TargetFramework $package_framework -TempPath $PackageData.TempPath -Offline
                         } Else {
-                            Import-Package $_.Name -Version $_.Version -TargetFramework $package_framework
+                            Import-Package $_.Name -Version $_.Version -TargetFramework $package_framework -TempPath $PackageData.TempPath
                         }
                         Write-Verbose "[Import-Package:Dependency-Handling] ($($PackageData.Name) Dependency) $($_.Name) Loaded"
                     }
@@ -357,9 +371,9 @@ function Import-Package {
                     } Else {
                         Write-Verbose "[Import-Package:Dependency-Handling] ($($PackageData.Name) Dependency) Loading $($_.Name) - $($_.Version) (Framework $( ([NuGet.Frameworks.NuGetFramework]$package_framework).GetShortFolderName() ))"
                         If( $PackageData.Offline ){
-                            Import-Package $_.Name -Version $_.Version -TargetFramework $package_framework -Offline
+                            Import-Package $_.Name -Version $_.Version -TargetFramework $package_framework -TempPath $PackageData.TempPath -Offline
                         } Else {
-                            Import-Package $_.Name -Version $_.Version -TargetFramework $package_framework
+                            Import-Package $_.Name -Version $_.Version -TargetFramework $package_framework -TempPath $PackageData.TempPath
                         }
                         Write-Verbose "[Import-Package:Dependency-Handling] ($($PackageData.Name) Dependency) $($_.Name) Loaded"
                     }
@@ -483,10 +497,12 @@ function Import-Package {
             Write-Warning "[Import-Package:Loading] $($PackageData.Name) is not needed for $( $bootstrapper.Runtime )`:$($TargetFramework.GetShortFolderName())"
         }
 
-        If( Test-Path (Join-Path $TempPath "*") ){
-            Write-Verbose "[Import-Package:Loading] Temp files: $TempPath"
-        } Else {
-            Remove-Item -Path $TempPath -ErrorAction Stop
+        If( $temp_path_generated ){
+            If( Test-Path (Join-Path $TempPath "*") ){
+                Write-Verbose "[Import-Package:Loading] Temp files: $TempPath"
+            } Else {
+                Remove-Item -Path $TempPath -ErrorAction Stop
+            }
         }
     }
 }
